+++ src/basic/ui/index.ts
@@ -1,4 +1,7 @@
 export * from './notification-item';
-export * from './search-input';
 export * from './admin-dashboard';
 export * from './product-list';
+export * from './cart';
+export * from './coupons';
+export * from './payments';
+export * from './header';
